requirements.txt
Flask
requests
Flask-OAuthlib
mysql.connector

